Move to number_symbolswrapper.cpp the SymbolsWrapper code from number_usageprefs.cpp
