+++ icu4c/source/i18n/i18n.vcxproj.filters
@@ -591,6 +591,9 @@
     <ClCompile Include="formatted_string_builder.cpp">
       <Filter>formatting</Filter>
     </ClCompile>
+    <ClCompile Include="number_usageprefs.cpp">
+      <Filter>formatting</Filter>
+    </ClCompile>
     <ClCompile Include="number_utils.cpp">
       <Filter>formatting</Filter>
     </ClCompile>
@@ -604,6 +607,9 @@
       <Filter>formatting</Filter>
     </ClCompile>
     <ClCompile Include="number_skeletons.cpp">
+      <Filter>formatting</Filter>
+    </ClCompile>
+    <ClCompile Include="number_symbolswrapper.cpp">
       <Filter>formatting</Filter>
     </ClCompile>
     <ClCompile Include="number_capi.cpp">
@@ -917,6 +923,9 @@
     <ClInclude Include="number_utypes.h">
       <Filter>formatting</Filter>
     </ClInclude>
+    <ClInclude Include="number_usageprefs.h">
+      <Filter>formatting</Filter>
+    </ClInclude>
     <ClInclude Include="number_utils.h">
       <Filter>formatting</Filter>
     </ClInclude>
@@ -930,6 +939,9 @@
       <Filter>formatting</Filter>
     </ClInclude>
     <ClInclude Include="number_skeletons.h">
+      <Filter>formatting</Filter>
+    </ClInclude>
+    <ClInclude Include="number_symbolswrapper.h">
       <Filter>formatting</Filter>
     </ClInclude>
     <ClInclude Include="string_segment.h">
