+++ icu4c/source/i18n/sources.txt
@@ -123,6 +123,7 @@ number_patternstring.cpp
 number_rounding.cpp
 number_scientific.cpp
 number_skeletons.cpp
+number_symbolswrapper.cpp
 number_usageprefs.cpp
 number_utils.cpp
 numfmt.cpp
